refactored simple/date reader to use same logic
