+++ asset/src/simple_api_reader/interfaces.ts
@@ -1,7 +1,1 @@
-import { ESReaderConfig } from '../elasticsearch_reader/interfaces';
 
-export interface ApiConfig extends ESReaderConfig {
-    endpoint: string;
-    token: string;
-    timeout: number;
-}
